de_project/requirements.txt
pandas>=2.0.0
pyarrow>=14.0.0
openpyxl>=3.1.0           # for .xlsx support
apache-airflow>=2.8.0
azure-storage-blob>=12.19.0